Fix lambda function naming bug with mount-ring!
There was a problem when trying to mount a ring handler using the `mount-ring!`
macro, where every handler you tried to mount would be uploaded to AWS with a
name of _I. The problem was that the mount-ring-fn was calculating the lambda
name assuming that the handler was coming in as a var, but the `mount-ring!` macro
was passing in the deref'd function instead, so the `(meta handler)` call was
always returning nil.

I moved the `make-lambda-function-name~ call to the ~mount-ring!~ macro and made
the ~mount-ring-fn~ accept the ~lambda-function-name~ as an option. This matches
the way that the non-ring ~mount!~ macro works.